+++ src/aircompressor/io/airlift/compress/zstd/HuffmanTableWriterWorkspace.java
@@ -24,6 +24,6 @@ class HuffmanTableWriterWorkspace
 
     // for compressing weights
     public final int[] counts = new int[MAX_TABLE_LOG + 1];
-    public final int[] normalizedCounts = new int[MAX_TABLE_LOG + 1];
+    public final short[] normalizedCounts = new short[MAX_TABLE_LOG + 1];
     public final FseCompressionTable fseTable = new FseCompressionTable(MAX_FSE_TABLE_LOG, MAX_TABLE_LOG);
 }
